refactor result tests
