+++ tests/Doctrine/ORM/Fixture/KitchenSinkRepository.php
@@ -4,6 +4,7 @@ namespace Zenstruck\Collection\Tests\Doctrine\ORM\Fixture;
 
 use Doctrine\ORM\EntityManagerInterface;
 use Zenstruck\Collection;
+use Zenstruck\Collection\Doctrine\ORM\EntityRepository;
 use Zenstruck\Collection\Doctrine\ORM\Repository\IsCollection;
 use Zenstruck\Collection\Doctrine\ORM\Repository\IsFilterable;
 use Zenstruck\Collection\Filterable;
@@ -13,7 +14,7 @@ use Zenstruck\Collection\Tests\Doctrine\Fixture\Entity;
 /**
  * @author Kevin Bond <kevinbond@gmail.com>
  */
-final class KitchenSinkRepository extends Collection\Doctrine\ORM\EntityRepository implements Filterable, Collection
+final class KitchenSinkRepository extends EntityRepository implements Filterable, Collection
 {
     use IsCollection, IsFilterable, Paginatable;
