+++ tests/Doctrine/ORM/Result/QueryFieldsResultTest.php
@@ -2,19 +2,15 @@
 
 namespace Zenstruck\Collection\Tests\Doctrine\ORM\Result;
 
-use PHPUnit\Framework\TestCase;
 use Zenstruck\Collection\Doctrine\ORM\Result;
 use Zenstruck\Collection\Tests\Doctrine\Fixture\Entity;
-use Zenstruck\Collection\Tests\Doctrine\HasDatabase;
-use Zenstruck\Collection\Tests\PagintableCollectionTests;
+use Zenstruck\Collection\Tests\Doctrine\ORM\ResultTest;
 
 /**
  * @author Kevin Bond <kevinbond@gmail.com>
  */
-final class QueryFieldsResultTest extends TestCase
+final class QueryFieldsResultTest extends ResultTest
 {
-    use HasDatabase, PagintableCollectionTests;
-
     /**
      * @test
      */
